Whitelist Storage Provider Test
Create a new class WhitelistStorageProviderImplTest, look for test cases that make use of Whitelist related methods in BridgeStorageProviderTest and move those tests to the new class.
